feat(game): add swipe controls
