+++ src/hooks/useSwipe.ts
@@ -1,0 +1,40 @@
+import { useEffect, useRef } from 'react';
+
+interface UseSwipeOptions {
+  onSwipeLeft?: () => void;
+  onSwipeRight?: () => void;
+  /** Minimum horizontal distance in pixels to trigger a swipe */
+  threshold?: number;
+}
+
+export const useSwipe = ({
+  onSwipeLeft,
+  onSwipeRight,
+  threshold = 30,
+}: UseSwipeOptions) => {
+  const startX = useRef<number | null>(null);
+
+  useEffect(() => {
+    const handleTouchStart = (e: TouchEvent) => {
+      startX.current = e.touches[0].clientX;
+    };
+
+    const handleTouchEnd = (e: TouchEvent) => {
+      if (startX.current === null) return;
+      const diff = e.changedTouches[0].clientX - startX.current;
+      if (diff > threshold) {
+        onSwipeRight?.();
+      } else if (diff < -threshold) {
+        onSwipeLeft?.();
+      }
+      startX.current = null;
+    };
+
+    window.addEventListener('touchstart', handleTouchStart);
+    window.addEventListener('touchend', handleTouchEnd);
+    return () => {
+      window.removeEventListener('touchstart', handleTouchStart);
+      window.removeEventListener('touchend', handleTouchEnd);
+    };
+  }, [onSwipeLeft, onSwipeRight, threshold]);
+};
